Bug fixed, add category cache

+++ server/src/main/java/cz/cvut/fel/nss/budgetmanager/BudgetManager/service/CategoryService.java
@@ -1,9 +1,13 @@
 package cz.cvut.fel.nss.budgetmanager.BudgetManager.service;
 
+import cz.cvut.fel.nss.budgetmanager.BudgetManager.exceptions.NotFoundException;
 import cz.cvut.fel.nss.budgetmanager.BudgetManager.model.Category;
+import cz.cvut.fel.nss.budgetmanager.BudgetManager.model.Transaction;
 import cz.cvut.fel.nss.budgetmanager.BudgetManager.repository.CategoryDao;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -29,12 +33,19 @@ public class CategoryService {
     }
 
     @Transactional
-    public void updateCategory(Category category) {
+    @CachePut(value = "categories", key = "#category.getCategoryId()")
+    public Category updateCategory(Category category) {
         categoryDao.update(category);
+        return category;
     }
 
     @Transactional
-    public void deleteCategory(Category category) {
+    @CacheEvict(value = "categories", key = "#id")
+    public void deleteCategory(Long id) {
+        Category category = categoryDao.find(id);
+        if (category == null) {
+            throw new NotFoundException("Category with id " + id + " was not found");
+        }
         categoryDao.remove(category);
     }
